Add joining_snap.sh

+++ excercise1/joining_output.sh
@@ -1,8 +1,17 @@
 #!/bin/bash
-cat output/out_static_00* > output/out_static.pgm
-rm output/out_static_00*
 
-cat output/out_ord_00* > output/out_ord.pgm
-rm output/out_ord_00*
+var=$(ls output | grep static_00 | wc -l)
+if [ ${var} -gt 0 ]
+then
+	cat output/out_static_00* > output/out_static.pgm
+	rm output/out_static_00*
+fi
+
+var=$(ls output | grep ord_00 | wc -l)
+if [ ${var} -gt 0 ]
+then
+	cat output/out_ord_00* > output/out_ord.pgm
+	rm output/out_ord_00*
+fi
